+++ backend/scoreMaker.ts
@@ -6,7 +6,7 @@ import {MusicNotation} from "@k-l-lambda/web-widgets";
 import {MIDI} from "@k-l-lambda/web-widgets";
 
 import {xml2ly, engraveSvg} from "./lilyCommands";
-import {LilyDocument} from "../inc/lilyParser";
+import {LilyDocument, replaceSourceToken} from "../inc/lilyParser";
 import * as staffSvg from "../inc/staffSvg";
 import {SingleLock} from "../inc/mutex";
 // eslint-disable-next-line
@@ -74,7 +74,7 @@ const xmlBufferToLy = async (xml: Buffer, options: LilyProcessOptions = {}): Pro
 const unescapeStringExp = exp => exp && exp.toString();
 
 
-const markScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON> => {
+const makeScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON> => {
 	const t0 = Date.now();
 
 	const engraving = await engraveSvg(source);
@@ -116,7 +116,7 @@ const markScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, log
 		const coverage = ((matcher.criterion.notes.length - omitC) / matcher.criterion.notes.length)
 			* ((matcher.sample.notes.length - omitS) / matcher.sample.notes.length);
 
-		logger.append("markScore.match", {coverage, omitC, omitS, path: matcher.path});
+		logger.append("makeScore.match", {coverage, omitC, omitS, path: matcher.path});
 	}
 
 	const matchedIds: Set<string> = new Set();
@@ -151,7 +151,14 @@ interface IncompleteScoreJSON {
 };
 
 
-const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON | IncompleteScoreJSON> => {
+interface SheetNotationResult extends IncompleteScoreJSON {
+	midiNotation: MusicNotation.NotationData;
+	sheetNotation: staffSvg.StaffNotation.SheetNotation;
+	lilyDocument: LilyDocument;
+};
+
+
+const makeScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON | IncompleteScoreJSON> => {
 	let midiNotation = null;
 
 	const pages = [];
@@ -271,14 +278,165 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 };
 
 
-const markScore = markScoreV2;
+const makeSheetNotation = async (source: string, lilyParser: GrammarParser, {withNotation = true, logger, lilyDocument}: {withNotation?: boolean, logger?: LogRecorder, lilyDocument?: LilyDocument} = {}): Promise<SheetNotationResult> => {
+	let midi = null;
+	let midiNotation = null;
+
+	const pages = [];
+	const hashTable = {};
+
+	const t0 = Date.now();
+
+	const attrGen = new SingleLock<LilyDocumentAttributeReadOnly>(true);
+
+	const engraving = await engraveSvg(source, {
+		// do some work during lilypond process running to save time
+		onProcStart: () => {
+			//console.log("tp.0:", Date.now() - t0);
+			if (!lilyDocument)
+				lilyDocument = new LilyDocument(lilyParser.parse(source));
+			attrGen.release(lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly);
+			//console.log("tp.1:", Date.now() - t0);
+		},
+		onMidiRead: withNotation && (midi_ => {
+			//console.log("tm.0:", Date.now() - t0);
+			midi = midi_;
+			midiNotation = midi && MusicNotation.Notation.parseMidi(midi);
+			//console.log("tm.1:", Date.now() - t0);
+		}),
+		onSvgRead: async svg => {
+			//console.log("ts.0:", Date.now() - t0);
+			const attributes = await attrGen.wait();
+			const page = staffSvg.parseSvgPage(svg, source, {DOMParser, logger, attributes});
+			pages.push(page.structure);
+			Object.assign(hashTable, page.hashTable);
+			//console.log("ts.1:", Date.now() - t0);
+		},
+	});
+
+	logger.append("scoreMaker.profile.engraving", {cost: Date.now() - t0});
+	logger.append("lilypond.log", engraving.logs);
+
+	const doc = new staffSvg.SheetDocument({pages});
+
+	const attributes = await attrGen.wait();
+	const meta = {
+		title: unescapeStringExp(attributes.title),
+		composer: unescapeStringExp(attributes.composer),
+		pageSize: doc.pageSize,
+		pageCount: doc.pages.length,
+		staffSize: attributes.staffSize,
+	};
+
+	const sheetNotation = staffSvg.StaffNotation.parseNotationFromSheetDocument(doc, {logger});
+
+	return {
+		midi,
+		midiNotation,
+		sheetNotation,
+		meta,
+		doc,
+		hashTable,
+		lilyDocument,
+	};
+};
+
+
+const makeScoreV3 = async (source: string, lilyParser: GrammarParser, {midi, logger, unfoldRepeats = false}: {midi?: MIDI.MidiData, logger?: LogRecorder, unfoldRepeats?: boolean} = {}): Promise<ScoreJSON | IncompleteScoreJSON> => {
+	const t0 = Date.now();
+
+	let lilyDocument = null;
+	let unfoldSource = null;
+
+	if (unfoldRepeats) {
+		lilyDocument = new LilyDocument(lilyParser.parse(source));
+		if (lilyDocument.containsRepeat()) {
+			lilyDocument.unfoldRepeats();
+			unfoldSource = lilyDocument.toString();
+
+			// keep 2 version lilypond source note href uniform
+			source = replaceSourceToken(unfoldSource, "\\unfoldRepeats");
+		}
+	}
+
+	const foldData = await makeSheetNotation(source, lilyParser, {logger, lilyDocument, withNotation: !midi && !unfoldSource});
+	const {meta, doc, hashTable} = foldData;
+
+	lilyDocument = lilyDocument || foldData.lilyDocument;
+	let midiNotation = foldData.midiNotation;
+	let sheetNotation = foldData.sheetNotation;
+
+	if (midi)
+		midiNotation = MusicNotation.Notation.parseMidi(midi);
+
+	if (unfoldSource) {
+		const unfoldData = await makeSheetNotation(unfoldSource, lilyParser, {logger, lilyDocument, withNotation: !midi});
+
+		midi = midi || unfoldData.midi;
+		midiNotation = unfoldData.midiNotation;
+		sheetNotation = unfoldData.sheetNotation;
+	}
+
+	midi = midi || foldData.midi;
+	if (!midi) {
+		console.warn("Neither lilypond or external arguments did not offer MIDI data, score maker finish incompletely.");
+		return {
+			meta,
+			doc,
+			midi,
+			hashTable,
+		};
+	}
+
+	const t5 = Date.now();
+
+	const matcher = await staffSvg.StaffNotation.matchNotations(midiNotation, sheetNotation);
+
+	logger.append("scoreMaker.profile.matching", {cost: Date.now() - t5});
+
+	if (logger && logger.enabled) {
+		const cis = new Set(Array(matcher.criterion.notes.length).keys());
+		matcher.path.forEach(ci => cis.delete(ci));
+
+		const omitC = cis.size;
+		const omitS = matcher.path.filter(ci => ci < 0).length;
+
+		const coverage = ((matcher.criterion.notes.length - omitC) / matcher.criterion.notes.length)
+			* ((matcher.sample.notes.length - omitS) / matcher.sample.notes.length);
+
+		logger.append("makeScore.match", {coverage, omitC, omitS, path: matcher.path});
+	}
+
+	const matchedIds: Set<string> = new Set();
+	midiNotation.notes.forEach(note => note.ids && note.ids.forEach(id => matchedIds.add(id)));
+
+	doc.updateMatchedTokens(matchedIds);
+
+	const pitchContextGroup = staffSvg.StaffNotation.createPitchContextGroup(sheetNotation.pitchContexts, midiNotation);
+
+	const noteLinkings = midiNotation.notes.map(note => _.pick(note, ["ids", "staffTrack", "contextIndex"]));
+
+	logger.append("scoreMaker.profile.full", {cost: Date.now() - t0});
+
+	return {
+		meta,
+		doc,
+		midi,
+		hashTable,
+		noteLinkings,
+		pitchContextGroup,
+	};
+};
+
+
+void makeScoreV1;
+void makeScoreV2;
+const makeScore = makeScoreV3;
 
 
 
 export {
 	markupLily,
 	xmlBufferToLy,
-	markScore,
-	markScoreV1,
-	markScoreV2,
+	makeScore,
 };